Move plugin_loader.py in scripts directory
